150 make permission for events

+++ app/Traits/DeleteEvent.php
@@ -3,41 +3,43 @@
 namespace App\Traits;
 
 use App\Models\Event;
-use App\Models\Folder;
 use Illuminate\Support\Facades\Storage;
 use Livewire\Attributes\On;
 
 trait DeleteEvent
 {
-    use HandleFolder;
-
     #[On('delete-event')]
     public function deleteEvent(int $id): void
     {
-        $event = Event::findOrFail($id)->load(['folders', 'folders.files']);
+        $this->authorize('delete', Event::class);
 
-        $folders = $event->folders;
-        $tasks = $event->tasks;
+        $event = Event::findOrFail($id)->load(['folders', 'folders.files', 'tasks']);
 
-        foreach ($folders as $folder) {
-            if ($folder->files) {
-                $disk = config('filesystems.default');
-                $original_path = config('events.original_path') . '/' . $folder->path;
+        $folders = $event->folders ?? null;
+        $tasks = $event->tasks ?? null;
 
-                foreach ($folder->files as $file) {
-                    $path = $original_path . '/' . $file->path;
+        if ($folders) {
+            foreach ($folders as $folder) {
+                if ($folder->files) {
+                    $disk = config('filesystems.default');
+                    $original_path = config('events.original_path') . '/' . $folder->path;
 
-                    if (Storage::disk($disk)->exists($path)) {
-                        Storage::disk($disk)->delete($path);
+                    foreach ($folder->files as $file) {
+                        $path = $original_path . '/' . $file->path;
+
+                        if (Storage::disk($disk)->exists($path)) {
+                            Storage::disk($disk)->delete($path);
+                        }
                     }
                 }
+                $folder->delete();
             }
-
-            $folder->delete();
         }
 
-        foreach ($tasks as $task) {
-            $task->delete();
+        if ($tasks) {
+            foreach ($tasks as $task) {
+                $task->delete();
+            }
         }
 
         $event->delete();
